requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# Anthropic AI (Opus 4.5)
anthropic==0.39.0

# Image Processing
pillow==11.0.0
numpy==2.1.3

# Utilities
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1

# Development
pytest==8.3.3
httpx==0.27.2
